requirements.txt
python-dotenv==1.0.0
zksync2==0.5.0
eth-account==0.9.0